[Transform] Extract common test code to TransformCommonRestTestCase class (#107103)
